% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_functions.R
\name{extract_vdem}
\alias{extract_vdem}
\title{Extracts parts of the V-Dem dataset by section number, name pattern, or label
pattern}
\usage{
extract_vdem(section_number = NULL, name_pattern = NULL,
  label_pattern = NULL, include_uncertainty = TRUE, include_sd = FALSE,
  include_osp = FALSE, include_ord = FALSE, include_nr = FALSE,
  include_external = FALSE)
}
\arguments{
\item{section_number}{A section number (or vector of section numbers) in the
V-dem codebook. Must be an integer or an integer vector (e.g., \code{1:2}, or
\code{28}).}

\item{name_pattern}{A character pattern, interpreted as a regular expression,
to search over the V-dem variable names, e.g., "v2x" extracts all variables
including the string "v2x" in their names.}

\item{label_pattern}{A character pattern, interpreted as a regular
expression, to search over the V-dem labels, e.g., "corrupt" extracts
all variables that mention "corrupt" in their label.}

\item{include_uncertainty}{Whether to return the uncertainty measurements for
the selected variables (\code{codelow} and \code{codehigh} suffixes). Default is
\code{TRUE}.}

\item{include_sd}{Whether to return the standard deviation for the selected
variables (\code{sd} suffix). Default is \code{FALSE}.}

\item{include_osp}{Whether to return variables available in the original
scale (\code{osp} suffix, linearized original scale posterior prediction).
Default is \code{FALSE}.}

\item{include_ord}{Whether to return variables available in an ordinal scale
(\code{ord} suffix, measurement model estimates of original scale value).
Default is \code{FALSE}.}

\item{include_nr}{Whether to return the number of coders for the variable
(\code{nr} suffix). Default is \code{FALSE}.}

\item{include_external}{Whether to return variables marked "external", i.e.,
coming from an external source rather than V-Dem (\code{e_} prefix). Default is
\code{FALSE}.}
}
\value{
A \link{tibble} with the V-Dem country identifiers, other state system
identifiers, plus the desired variables.
}
\description{
Extracts variables in the V-Dem dataset (version 7.1), according to the
section number in the codebook or a regular expression pattern matching the
name or the label.
}
\examples{
# Main democracy indexes
extract_vdem(section_number = 1)
# Mid-level indexes, no measures of uncertainty
extract_vdem(section_number = 2, include_uncertainty = FALSE)
# All indexes about corruption
extract_vdem(label_pattern = "corrupt")
# All indexes about elections, include external indexes
extract_vdem(label_pattern = "elect", include_external = TRUE)
# All civil liberties indexes, include original scale prediction
extract_vdem(name_pattern = "^v2cl", include_osp = TRUE)
# Combine section and label patterns
extract_vdem(section_number = 1, label_pattern = "elect", include_external = TRUE)
}
