% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_functions.R
\name{extract_vdem_filter}
\alias{extract_vdem_filter}
\title{Extracts parts of the V-Dem dataset by passing a filter expression to
\link{vdem_codebook}}
\usage{
extract_vdem_filter(
  ...,
  include_uncertainty = TRUE,
  include_sd = FALSE,
  include_osp = FALSE,
  include_ord = FALSE,
  include_nr = FALSE,
  include_external = FALSE
)
}
\arguments{
\item{...}{A sequence of filter expressions, separated by commas, suitable
for use with \link[dplyr:filter]{dplyr::filter}, operating on the column names of
\link{vdem_codebook}. See examples.}

\item{include_uncertainty}{Whether to return the uncertainty measurements for
the selected variables (\code{codelow} and \code{codehigh} suffixes). Default is
\code{TRUE}.}

\item{include_sd}{Whether to return the standard deviation for the selected
variables (\code{sd} suffix). Default is \code{FALSE}.}

\item{include_osp}{Whether to return variables available in the original
scale (\code{osp} suffix, linearized original scale posterior prediction).
Default is \code{FALSE}.}

\item{include_ord}{Whether to return variables available in an ordinal scale
(\code{ord} suffix, measurement model estimates of original scale value).
Default is \code{FALSE}.}

\item{include_nr}{Whether to return the number of coders for the variable
(\code{nr} suffix). Default is \code{FALSE}.}

\item{include_external}{Whether to return variables marked "external", i.e.,
coming from an external source rather than V-Dem (\code{e_} prefix). Default is
\code{FALSE}.}
}
\value{
A \link{tibble} with the V-Dem country identifiers, other state system
identifiers, plus the desired variables.
}
\description{
Extracts variables in the V-Dem dataset (version 11), by using a filter
expression that is used to extract variable names from the V-Dem coddebook
(\link{vdem_codebook})
}
\examples{
# All in section 2 of the codebook except component indexes
extract_vdem_filter(section == 2, !grepl("component index", label))
# All variables inclued in release 2 of the dataset
extract_vdem_filter(grepl("2(,|$)|1-11", data_release))
}
