% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_documentation.R
\docType{data}
\name{vdem_codebook}
\alias{vdem_codebook}
\title{V-Dem codebook}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1216 rows and 18 columns.}
\usage{
vdem_codebook
}
\description{
The V-Dem codebook (version 7.1) as a \link{tibble} for easier use in R. The original codebook is available
at \url{https://www.v-dem.net/en/data/data-version-7-1/}
}
\section{Variable Descriptions}{


\describe{

\item{section}{The section number in the codebook, as an integer. For
example, section 1 contains the main indexes of democracy, section 2 the
mid-level indexes, etc. Note section 28 is dubplicated in the codebook.}

\item{name}{The name of the variable. Does not include suffixes (e.g.,
\code{_codehigh}, \code{_osp})}

\item{label}{The label extracted from the codebook.}

\item{type}{The type of variable. These are, according to the codebook:
\itemize{
\item A*: Coding by Research Assistants. This is based on extant sources and is
factual in nature. Country Experts indicate their confidence for this
pre-coded data.
\item A: Coding by Project Managers and Research Assistants.
This is based on extant sources and is factual in nature.
\item B: Coding by
Country Coordinators. This is typically a graduate student or recent
graduate who is from the country. These questions are also factual in
nature.
\item C: Coding by Country Experts. This is typically a scholar or
professional with deep knowledge of a country and perhaps of a particular
political institution. Generally, that person is a citizen or resident of
the country being coded. Multiple experts (usually 5 or more) code each
variable.
\item D: Indices, i.e., composite variables composed of type A, B, or
C variables. This may be accomplished by adding a denominator (e.g., per
capita), by creating a cumulative scale (Total number of ...), or by
aggregating larger concepts (e.g., components or indices of democracy).
\item E: Non-V-Dem. If we import a variable from another source without doing any
original coding, except for perhaps imputing missing data, it is not
considered a V-Dem product. Many of these variables are found in the
section of the Codebook labeled "Background Factors.: If, however, we
gather data from a number of sources and combine them in a more than purely
mechanical fashion (requiring some judgment on our part), we regard this as
a V-Dem product and classify it as type A, B, C, or D. All "E" variables
are included in the codebook except those drawn from sources that have more
than 30 variables such as Archigos (Goemans et al.), BDM (Bueno de Mesquita
et al.), Henisz/POLCON (2000, 2002), Miller (Democratic Pieces),
Performance of Nations (Kugler and Tammen), PEI (Norris et al.), PIPE
(Przeworski et al.), QoG (Quality of Government). For these variables, we
ask users to consult separate codebooks, as listed above.}
}

\item{project_manager}{The project manager listed in the codebook.}

\item{compiler}{The compiler listed in the codebook.}

\item{question}{If available, the question answered by the coders.}

\item{clarification}{If available, the clarification in the codebook.}

\item{responses}{If available, the possible responses.}

\item{scale}{If available, the scale of the responses (e.g., nominal,
categorical).}

\item{aggregation}{If available, the type of aggregation for the index
(e.g., Bayesian factor analysis, average of indicators).}

\item{note}{If available, any notes in the codebook.}

\item{ordering}{If available, the order in which the variable is coded
(e.g., only if previous quesiton is answered).}

\item{cross_coder}{If available, the method for aggregating multiple
coders' judgments.}

\item{source}{If available, the source listed in the codebook.}

\item{data_release}{If available, the version of the V-Dem data where the
variable was included.}

\item{data_release}{If available, the suggested citation listed in the
codebook.}

}
}

\references{
Coppedge, Michael, John Gerring, Staffan I. Lindberg, Svend-Erik
Skaaning, Jan Teorell, David Altman, Frida Andersson, Michael Bernhard, M.
Steven Fish, Adam Glynn, Allen Hicken, Carl Henrik Knutsen, Kyle L.
Marquardt, Kelly McMann, Valeriya Mechkova, Pamela Paxton, Daniel Pemstein,
Laura Saxer, Brigitte Seim, Rachel Sigman and Jeffrey Staton. 2017. "V-Dem
Codebook v7.1" Varieties of Democracy (V-Dem) Project. Available at
\url{https://www.v-dem.net/en/data/data-version-7-1/}
}
\keyword{datasets}
