% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_documentation.R
\docType{data}
\name{VDem_plus}
\alias{VDem_plus}
\title{V-Dem dataset (version 7.1)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17604 rows and 3615 columns.}
\source{
Coppedge, Michael, John Gerring, Staffan I. Lindberg, Svend-Erik
Skaaning, Jan Teorell, David Altman, Michael Bernhard, M. Steven Fish, Adam
Glynn, Allen Hicken, Carl Henrik Knutsen, Joshua Krusell, Anna Luhrmann,
Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Moa Olin, Pamela
Paxton, Daniel Pemstein, Josefine Pernes, Constanza Sanhueza Petrarca,
Johannes von Romer, Laura Saxer, Brigitte Seim, Rachel Sigman, Jeffrey
Staton, Natalia Stepanova, and Steven Wilson. 2017. "V-Dem
\href{https://www.v-dem.net/en/data/data-version-7-1/}{Country-Year/Country-Date}
Dataset v7.1" Varieties of Democracy (V-Dem) Project.  Available at
\url{https://www.v-dem.net/en/data/data-version-7-1/}
}
\usage{
VDem_plus
}
\description{
The V-Dem dataset (version 7.1), including external variables. This
documentation only provides variable names and labels for the country-year
identifiers included. The rest of the codebook is available at
\url{https://www.v-dem.net/en/data/data-version-7-1/},
and in this package as \link{vdem_codebook}.
}
\section{Identifiers}{


\describe{

\item{vdem_country_name}{The original country name in the dataset. There
are 178 distinct countries in version 7.1. According to the codebook, "a
V-Dem country is a political unit enjoying at least some degree of
functional and/or formal sovereignty."}

\item{vdem_country_id}{Unique numerical country id for each country.}

\item{vdem_country_text_id}{Unique abbreviation for each country.}

\item{year}{Year, from 1900-2016.}

\item{extended_country_name}{The standard country according to the
\href{http://privatewww.essex.ac.uk/~ksg/statelist.html}{Gleditsch and Ward system ofstates} (Gleditsch and
Ward 1999), or a standardized country name. Note that West Bank, and Gaza
are two separate coding units in \code{vdem_country_name} but all map to
"Palestine, State of" here. Column added by Xavier Marquez to the original
V-Dem dataset.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch
and Ward list of independent states. This differs from the V-Dem COW code
in a few instances. For details, see Gleditsch and Ward 1999. The list can
be found at \url{http://privatewww.essex.ac.uk/~ksg/statelist.html}. Column
added by Xavier Marquez to the original V-Dem dataset.}

\item{historical_date}{According to the codebook, "This variable is
included in the V-Dem Country Date dataset. The default date is 31st
December, as in 2016-12-31, referring to the time span from 01-01 to 12-31
in a respective year. Additionally, specific changes, such as the
appointment of a Head of State, are coded on the specific date within a
certain year. Thus, a code can change within a year, and will be reflected
in the 12-31 date."}

\item{codingstart}{"V-Dem country coding starts in 1900, or from when a
country first was formally sovereign (according to international law). For
detailed information, please see to VDem Country Coding Units."}

\item{gapstart}{"Time periods when a country does not fulfill V-Dem's
coding period criteria are not coded. The date that indicates the gap start
is the last date coded before the gap. For more details about V-Dem country
coding periods, please see to V-Dem Country Coding Units."}

\item{gapend}{"The periods of when a country does not fulfill V-Dem's
coding period criteria are not coded. The date that indicates the gap end
is the first date coded after the gap. For more details about V-Dem country
coding periods, please see to V-Dem Country Coding Units."}

\item{codingend}{"V-Dem country coding ends in 2016, or from when a country
formally stopped being sovereign (according to international law). For
detailed information , please see VDem Country Coding Units."}

\item{vdem_cown}{The Correlates of War numeric country code. Note that this
differs from Gleditsch and Ward's numeric country code in a few cases.}

\item{cown}{The Correlates of War numeric country code (from the 2016
version of the COW state list). Note that this differs from Gleditsch and
Ward's numeric country code in a few cases; it also differs from
\code{vdem_cown} in a couple of cases: Germany 1949-1989, where \code{vdem_cown}
incorrectly has 255 (it should be 260), Czech Republic 1918-1992 (coded as
Czechoslovakia, 315 in \code{cown}, 316 in \code{vdem_cown}), South Korea 1900-1908
(coded as Korea, 730, where \code{vdem_cown} has 732), and Yemen 1918-1989
(coded as 678, where \code{vdem_cown} has 679). Column added by Xavier Marquez
to the original V-Dem dataset.}

\item{GWc}{Gleditsch and Ward's alphabetic country code, from the Gleditsch
and Ward list of independent states. For details, see Gleditsch and Ward
1999. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}. Column added by
Xavier Marquez to the original V-Dem dataset.}

\item{extended_region}{Region. Column added by Xavier Marquez to the original V-Dem
dataset.}

\item{extended_continent}{Continent. Column added by Xavier Marquez to the original
V-Dem dataset.}

\item{GW_startdate}{The entry date of the state into the international
system (usually the date of independence, though there can be more than
one), according to Gleditsch and Ward, or 1 January 1816, whichever is
later. Column added by Xavier Marquez to the original V-Dem dataset.}

\item{GW_enddate}{The exit date of the state from the international system
(usually the date the state lost its independence, though there can be more
than one), according to Gleditsch and Ward, or NA if the country is still
in the international system as of the time the dataset was compiled. Column
added by Xavier Marquez to the original V-Dem dataset.}

\item{microstate}{Whether the state is a microstate, according to
Gleditsch. His tentative list of microstates is available at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}. Column added by
Xavier Marquez to the original V-Dem dataset.}

\item{lat}{The rough latitude of the state. Column added by Xavier Marquez
to the original V-Dem dataset.}

\item{lon}{The rough longitude of the state. Column added by Xavier Marquez
to the original V-Dem dataset.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Column added by Xavier Marquez to the original V-Dem
dataset. You may instead wish to use one of the v2sv* columns provided in
the original V-Dem dataset.}

}
}

\references{
Coppedge, Michael, John Gerring, Staffan I. Lindberg, Svend-Erik Skaaning,
Jan Teorell, David Altman, Frida Andersson, Michael Bernhard, M. Steven Fish,
Adam Glynn, Allen Hicken, Carl Henrik Knutsen, Kyle L. Marquardt, Kelly
McMann, Valeriya Mechkova, Pamela Paxton, Daniel Pemstein, Laura Saxer,
Brigitte Seim, Rachel Sigman and Jeffrey Staton. 2017. "V-Dem Codebook v7.1"
Varieties of Democracy (V-Dem) Project. Available at
\url{https://www.v-dem.net/en/data/data-version-7-1/}

Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System
Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}
}
\keyword{datasets}
