% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_documentation.R
\docType{data}
\name{VDem_plus}
\alias{VDem_plus}
\title{V-Dem dataset (version 11)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27192 rows and 4188 columns.
}
\source{
Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I.
Lindberg, Jan Teorell, David Altman, Michael Bernhard, Agnes Cornell, M.
Steven Fish, Lisa Gastaldi, Haakon Gjerløw, Adam Glynn, Allen Hicken, Anna
Lührmann, Seraphine F. Maerz, Kyle L. Marquardt, Kelly McMann, Valeriya
Mechkova, Pamela Paxton, Daniel Pemstein, Johannes von Römer, Brigitte
Seim, Rachel Sigman, Svend-Erik Skaaning, Jeffrey Staton, Aksel Sundtr¨om,
Eitan Tzelgov,Luca Uberti, Yi-ting Wang, Tore Wig, and Daniel Ziblatt.
2021. "V-Dem Codebook v11" Varieties of Democracy (V-Dem) Project.
Available at
\url{https://www.v-dem.net/en/data/data-version-11/}
}
\usage{
VDem_plus
}
\description{
The V-Dem dataset (version 11), including external variables. This
documentation only provides variable names and labels for the country-year
identifiers included. The rest of the codebook is available at
\url{https://www.v-dem.net/en/data/data-version-11/},
and in this package as \link{vdem_codebook}.
}
\section{Identifiers}{


\describe{

\item{vdem_country_name}{The original country name in the dataset. There
are 202 distinct countries in version 11. According to the codebook, "a
V-Dem country is a political unit enjoying at least some degree of
functional and/or formal sovereignty."}

\item{vdem_country_id}{Unique numerical country id for each country.}

\item{vdem_country_text_id}{Unique abbreviation for each country.}

\item{year}{Year, from 1789-2020.}

\item{extended_country_name}{The standard country according to the
\href{http://privatewww.essex.ac.uk/~ksg/statelist.html}{Gleditsch and Ward system of states} (Gleditsch and
Ward 1999), or a standardized country name. Note that West Bank, and Gaza
are two separate coding units in \code{vdem_country_name} but all map to
"Palestine, State of" here. Column added by Xavier Marquez to the original
V-Dem dataset.}

\item{GWn}{Gleditsch and Ward's numeric country code, from the Gleditsch
and Ward list of independent states. This differs from the V-Dem COW code
in a few instances, and it does not exist in for some historical polities.
For details, see Gleditsch and Ward 1999. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}. Column added by
Xavier Marquez to the original V-Dem dataset.}

\item{historical_date}{"This variable is included in the V-Dem Country Date
dataset. The default date is December 31st, as in 2020-12-31, referring to
the time span from 01-01 to 12-31 in a respective year. Additionally,
specific changes, such as the appointment of a Head of State, are coded on
the specific date within a certain year. Thus, a code can change within a
year, and will be reflected in the 12-31 date."}

\item{codingstart}{"V-Dem country coding starts in 1789, or from when a
country first enjoyed at least some degree of functional and/or formal
sovereignty. For detailed information, please see the V-Dem Country Coding
Units document."}

\item{codingstart_contemp}{This variable indicates the coding start for the
countries coded by Contemporary V-Dem.}

\item{codingstart_hist}{This variable indicates the coding start for the
countries coded by Historical V-Dem.}

\item{gapstart}{"Time periods when a country does not fulfill V-Dem’s
coding period criteria are not coded. The date that indicates the gap start
is the last date coded before the gap. For more details about V-Dem country
coding periods, please see the V-Dem Country Coding Units document."}

\item{gapend}{"The periods of when a country does not fulfill V-Dem’s
coding period macriteria are not coded. The date that indicates the gap end
is the first date coded after the gap. For more details about V-Dem country
coding periods, please see the V-Dem Country Coding Units document."}

\item{codingend}{"V-Dem country coding ends in 2018, or from when a country
formally stopped enjoying at least some degree of functional and/or formal
sovereignty. For detailed information, please see the V-Dem Country Coding
Units document."}

\item{codingend_contemp}{"This variable indicates when the coding ends for
countries coded by the Contemporary V-Dem project."}

\item{codingend_hist}{"This variable indicates when the coding ends for
countries coded by the Historical V-Dem Project."}

\item{project}{"This variable indicates which V-Dem project coded that
country-year: Contemporary V-Dem (0), Historical V-Dem (1), or both (2)."}

\item{historical}{"This variable indicates if the Historical V-Dem project
coded a country at any time: No (0), Yes (1)."}

\item{vdem_cown}{The Correlates of War numeric country code (2016). Note
that this differs from Gleditsch and Ward's numeric country code in a few
cases.}

\item{cown}{The Correlates of War numeric country code (from the 2016
version of the COW state list). Note that this differs from Gleditsch and
Ward's numeric country code in a few cases. Column added by Xavier Marquez
to the original V-Dem dataset.}

\item{GWc}{Gleditsch and Ward's alphabetic country code, from the Gleditsch
and Ward list of independent states. Missing for some historical polities
and some states with contested sovereignty. For details, see Gleditsch and
Ward 1999. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}. Column added by
Xavier Marquez to the original V-Dem dataset.}

\item{extended_region}{Region. Column added by Xavier Marquez to the
original V-Dem dataset. You may wish to use the \verb{e_region*} set of
variables instead.}

\item{extended_continent}{Continent. Column added by Xavier Marquez to the
original V-Dem dataset.}

\item{GW_startdate}{The entry date of the state into the international
system (usually the date of independence, though there can be more than
one), according to Gleditsch and Ward, or 1 January 1816, whichever is
later. Column added by Xavier Marquez to the original V-Dem dataset.}

\item{GW_enddate}{The exit date of the state from the international system
(usually the date the state lost its independence, though there can be more
than one), according to Gleditsch and Ward, or NA if the country is still
in the international system as of the time the dataset was compiled. Column
added by Xavier Marquez to the original V-Dem dataset.}

\item{microstate}{Whether the state is a microstate, according to
Gleditsch. His tentative list of microstates is available at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}. Column added by
Xavier Marquez to the original V-Dem dataset.}

\item{lat}{The rough latitude of the state. Column added by Xavier Marquez
to the original V-Dem dataset.}

\item{lon}{The rough longitude of the state. Column added by Xavier Marquez
to the original V-Dem dataset.}

\item{in_GW_system}{Whether the state is "in system" (that is, is
independent and sovereign), according to Gleditsch and Ward, for this
particular date. Column added by Xavier Marquez to the original V-Dem
dataset. You may instead wish to use one of the `v2sv*`` columns provided
in the original V-Dem dataset.}

}
}

\references{
Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I. Lindberg,
Jan Teorell, Nazifa Alizada, David Altman, Michael Bernhard, Agnes Cornell,
M. Steven Fish, Lisa Gastaldi, Haakon Gjerløw, Adam Glynn, Allen Hicken,
Garry Hindle, Nina Ilchenko, Joshua Krusell, Anna Luhrmann, Seraphine F.
Maerz, Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Juraj Medzihorsky,
Pamela Paxton, Daniel Pemstein, Joseﬁne Pernes, Johannes von Römer, Brigitte
Seim, Rachel Sigman, Svend-Erik Skaaning, Jeffrey Staton, Aksel Sundström,
Ei-tan Tzelgov, Yi-ting Wang, Tore Wig, Steven Wilson and Daniel Ziblatt.
2021. "V-Dem Country-Year Dataset v11" Varieties of Democracy
Project. https://doi.org/10.23696/vdems21

Pemstein, Daniel, Kyle L. Marquardt, Eitan Tzelgov, Yi-ting Wang, Juraj
Medzihorsky, Joshua Krusell, Farhad Miri, and Johannes von Römer. 2020 "The
V-Dem Measurement Model: Latent Variable Analysis for Cross-National and
Cross-Temporal Expert-Coded Data". V-Dem Working Paper No. 21. 5th edition.
University of Gothenburg: Varieties of Democracy Institute.

Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System
Membership: A Revised List of the Independent States since 1816."
International Interactions 25: 393-413. The list can be found at
\url{http://privatewww.essex.ac.uk/~ksg/statelist.html}
}
\keyword{datasets}
